Restructure the admin folder in the pages directory
