+++ src/pages/admin/Communication.tsx
@@ -1,15 +1,15 @@
 import { Box, Toolbar, Typography } from "@mui/material";
 import React from "react";
-import AdminTaskbar from "../components/AdminTaskbar";
+import AdminTaskbar from "../../components/AdminTaskbar";
 
-const AdminPage = () => {
+const Communication = () => {
 	return (
 		<div>
 			<AdminTaskbar />
 			<Box component="main" sx={{ flexGrow: 1, p: 15 }}>
 				<Toolbar />
 				<Typography paragraph align="center">
-          Admin Page
+				Message Center
 				</Typography>
 			</Box>
 		</div>
@@ -17,4 +17,4 @@ const AdminPage = () => {
 	);
 };
 
-export default AdminPage;
+export default Communication;
