+++ src/pages/admin/Dashboard.tsx
@@ -1,15 +1,15 @@
 import { Box, Toolbar, Typography } from "@mui/material";
 import React from "react";
-import AdminTaskbar from "../components/AdminTaskbar";
+import AdminTaskbar from "../../components/AdminTaskbar";
 
-const AdminCommunication = () => {
+const Dashboard = () => {
 	return (
 		<div>
 			<AdminTaskbar />
 			<Box component="main" sx={{ flexGrow: 1, p: 15 }}>
 				<Toolbar />
 				<Typography paragraph align="center">
-				Message Center
+				2022 Blueberry Fundraiser - Dashboard
 				</Typography>
 			</Box>
 		</div>
@@ -17,4 +17,4 @@ const AdminCommunication = () => {
 	);
 };
 
-export default AdminCommunication;
+export default Dashboard;
