+++ src/pages/admin/Page.tsx
@@ -1,15 +1,15 @@
 import { Box, Toolbar, Typography } from "@mui/material";
 import React from "react";
-import AdminTaskbar from "../components/AdminTaskbar";
+import AdminTaskbar from "../../components/AdminTaskbar";
 
-const AdminDashboard = () => {
+const AdminPage = () => {
 	return (
 		<div>
 			<AdminTaskbar />
 			<Box component="main" sx={{ flexGrow: 1, p: 15 }}>
 				<Toolbar />
 				<Typography paragraph align="center">
-				2022 Blueberry Fundraiser - Dashboard
+          Admin Page
 				</Typography>
 			</Box>
 		</div>
@@ -17,4 +17,4 @@ const AdminDashboard = () => {
 	);
 };
 
-export default AdminDashboard;
+export default AdminPage;
